requirements.txt
aioredis==1.3.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.3
django-sslserver==0.22
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
Markdown==3.3.4
martor==1.6.12
msgpack==1.0.3
msgpack-python==0.5.6
packaging==21.3
psycopg==3.0.10
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-decouple==3.6
redis==2.10.6
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
zope.interface==5.4.0
